requirements.txt
streamlit
requests
pandas
fastapi
uvicorn
python-multipart
pymupdf
beautifulsoup4
chromadb
sentence-transformers
pydantic
python-dotenv
google-generativeai
openai
